framework-less/Live-API/RAG-Engine/audio2audio/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
google-cloud-aiplatform
numpy
sounddevice